churn: moved the domain resolving helper to config
